ci: Use more container tech
Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/735.

This commit continues the trend implemented in a prior CI upgrade: all
dependencies are "burned into" container images before the main body of
the pipeline begins. This removes hiccups that are caused by flaky
network and external servers, and makes the jobs far easier to
reproduce locally. And should be faster overall.

The new ci/containers/ directory contains the buildscripts for all the
images we use in CI, absorbing most of the usage of the CI images repo.
Note that the CI images are still used in some places, particularly the
Spack pipeline where no alternative images exist currently. Changing
that requires a handful of Spack PRs to (be written and) land first.