Reimplement ROCm kernels as binary-less kernels
We don't actually get any PC sampling data from ROCm currently, so we
really only need the kernel name. This commit removes the code that
saved and looked up functions in the ROCm binaries, instead opting for
the GPU-agnostic kernel name table.

This removes the need to analyze the ROCm binaries, which is suspect
due to Dyninst's current lack of complete support for the AMD ISAs.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/738.