ci: Disallow GPU access without proper tags
Apparently the CUDA Container Toolkit will allow access to the GPUs even
if the Runner didn't explicitly allow access, so the one CUDA-based
check job that was designed not to have GPU access still... did.
This is madness and very worrying from a security perspective.

For now, this makes the CI "self-regulating" in that it relinquishes
access to the GPUs for the job in question. This should help reduce some
of the CUPTI_HARDWARE_BUSY errors seen in CI.