Skip when required GPU(s) are not available
We want to sometimes run the test suite with +cuda but without any GPUs
actually available. This commit adds the pieces to hpctoolkit.test to
support that case.

To make life easier, cuda_filter is now reimplemented in Python using
data gathered from nvidia-smi. Something similar can/will be done for
HIP in the future.