dev: Absorb and update ci.buildfe as ./dev buildfe
The paradigm shift here is that ci.buildfe now modifies a valid
./configure argument set, instead of generating arguments from a
template. In other words, this means ./dev create generates a "default"
set of arguments, and then ci.buildfe modifies it to match the spec.

As a side effect, this means ci.buildfe knows a bit less about the
dependencies of HPCToolkit. Eventually this will be nothing (i.e. all
flags will be "feature"-type --enable/--disable flags), but we're not
quite there yet.

In the meantime, the ./dev scriptsuite uses a proper Python packaging
scheme and is easier to set up than ci.buildfe. Pulling ci.buildfe into
the ./dev scriptsuite makes the whole cleaner and opens up a range of
future possibilities.

There are many things I would love to take time and improve in what is
now dev.buildfe, but this does the job for now.