Update the CI to use the ./dev scriptsuite
So instead of concretizing a previously present environment, it builds
and stores a "fresh" one generated by ./dev generate. Otherwise there
isn't much difference, except now the images have the ./dev scriptsuite
installed which is used in the CI (and for developer convenience).

Also update the ci.buildfe calls to use the new ./dev buildfe paradigm.

Drop ext-libs/liblzma.so from the install manifest. It's not installed
if xz was compiled without `libs=shared`, and it's not critical for the
operation of hpcrun as a whole.