tests2: Don't capture stderr from nvidia-smi
If we run nvidia-smi and it fails for one reason or another, we fail the
test. But without the stderr from nvidia-smi there's no way to know
why it failed. Fix that by redirecting only stdout and not stderr from
the subprocess.

Also remove a debug print() that was left in by accident.