ci/containers: Enable Buildah caching
Building the container images used in CI takes a long while. Most of the
tasks are identical each time and so could in theory be cached with high
hit rates.

This commit enables buildah layer caching, and makes adjustments to how
the Spack local binary cache is presented inside the container to dodge
https://github.com/containers/buildah/issues/4647.