ci: Use require to enforce generic architectures
`require:` is a stronger enforcement than `target:` in Spack, so we
can use the former instead. This allows us to remove the determination
of the architecture to use, instead allowing Spack to implicitly
determine the architecture from the running container's arch.

This is a tighter limit than was used before, because modern Spack has
forgotten what `target=cortex_a75` is.