ci: Remove compilers.yaml from the images
We aren't using Spack (heavily) in these images, all we use it for is to
install from the buildcache. Spack doesn't need to know about *all* the
compilers we installed to do that much.

This also fixes an issue regarding the Spack code for OpenSUSE Leap 15,
it used to be called `os=opensuse_leap15` but now is `os=opensuse15`.
Bisection reveals the difference appeared when the distro vendered
dependency was updated in https://github.com/spack/spack/pull/33951 to
include https://github.com/python-distro/distro/pull/319.

The change in Spack occured in January 2023, years after the change in
distro was merged in October 2021, over a year after OpenSUSE Leap 15.2
was released in July 2020. Sigh.