ci: Use a single shared buildcache
Up to this point, the Spack buildcache used in CI has been split by
MR/branch, following the pattern used by the CI of Spack upstream. It's
a critical security feature for them to ensure the packages of one PR
don't taint other PRs and their public buildcache.

It turns out this consideration is unnecessary for us: if an MR comes
from a fork GitLab will run its CI using the fork's secrets, not ours
(unless given permission by a Developer). This means our private
buildcache can only be used with our permission, so we don't need the
complicated setup Spack's CI needs.

Decomplicate the CI by using a single private buildcache shared between
all branches and MRs. The configuration for this buildcache is now in a
single CI variable SBCACHE_YAML, in the format of Spack's mirrors.yaml.