ci: Use a Spack environment for Autotools
The Autotools versions haven't changed in a long while, and we will
eventually move to Meson as per https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/710.

Until that happens though, we can simplify the pipeline a bit by using
the deps image(s) in the autoreconf jobs instead of an out-of-band image.
The name for this new environment is `_autogen` to match with
https://gitlab.com/hpctoolkit/hpctoolkit/-/merge_requests/806.