tests2: Fix incorrect algorithm for flex-fields
Context nodes are represented in the meta.db using "flex-fields," a
compact way to represent a series of optional fields without sacrificing
the required alignments. Unfortunately the code used by the test suite
did not correctly implement the logic, causing tests to crash in certain
never-before-seen cases.

This commit rewrites the algorithm to be simpler and correct, at the
cost of efficiency.