ci: Drop autoreconf check
This linting job checks that every commit in an MR has had its autogoo
regenerated. Unfortunately its rules: configuration sometimes causes
the entire pipeline to fail to fire, as it does in this branch.

I believe this is a GitLab bug, but it is also a heavy-handed job with
rather limited utility in practice. So, dropping it for now to unblock
this MR.