tests2: Make the test utilities a Python package
The test utilities are exposed to the test scripts via PYTHONPATH, but
also require a few dependencies of their own (click, pyparsing, etc.).
Unfortunately, if said dependencies are supplied via PYTHONPATH they get
lost due to a quirk of Meson's.

Smooth out this process by making the test utilities a Python package,
which gets installed in a venv at configure time. The dependencies
aren't installed, but pip will still check that they are installed and
of appropriate versions. Using a venv avoids the need for PYTHONPATH.

To avoid similar issues while trying to prepend elements to the PATH,
the test scripts are transformed at configure time to burn in the full
path to the venv Python.