struct: add support for dyninst's unknown instruction API
Add a callback function to catch Dyninst's unknown instruction API,
use XED to find the length of the next instruction and return a fake
no-op for InstructionAPI.  This gives a way to avoid large unclaimed
regions of code (gaps) on sapphire rapids.

Test for Dyninst support with configure.

Add debug output to summarize the remaining gaps.

For now, this only works with x86 with XED.