hpcprof: Don't crash on error copying source files
The source files aren't a critical part of a performance database, the
end result will still be perfectly usable. So if we fail to copy a
source file into the database for any reason, catch the error and make
it a WARNING instead.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/761