dev: Validate spec versions on load
When loading packages from a devenv, ensure that the versions of the
packages loaded match those that were expected from the outset, rather
than blindly assuming the Spack environment is consistent.

This acts as the "check" to ensure that the reused deps image is valid,
otherwise the script will simply crash and tell the user to update.

To ensure this failure happens early in a pipeline, a new validation job
is added which does nothing more than trigger these checks.