hpcprof-mpi: Fix/hack inconsistent Metric partials
The Prof-MPI (distributed-memory) code path makes a number of
assumptions about the "shared" state. One of these is that the
StatisticPartials registered for a Metric are consistent across all
MPI ranks. Unfortunately they are not always, an issue which only seems
to appear with the additional changes introduced here.

The main offender seems to be "invisible" Metrics, which do not have
Statistics but do have a "sum" Partial. So for now, always generate the
"sum" Partial even if it is not actually required. It's a hack but it
works until a better distributed-memory algorithm can be implemented.

While we're here, add additional consistency checks when unpacking the
Metrics so that issues like this can't happen again.