dev.buildfe: Identify warnings by line, not lineno
GitLab will report how many code quality issues were created and fixed
by any particular MR, using the "fingerprint" hashes written in each
report JSON object as unique identifiers. Previously, for compiler
warnings this "fingerprint" was simply the hash of the report, which
includes the line number(s) of the warning. This means that many MRs
changing a file that contains warnings will get a warning widget saying
that said warnings were both "created and fixed" (with differing line
numbers).

Adjust the fingerprints to use the actual line of code instead of the
line number. In theory, this will make it so MRs that change the
affected line will get a flag (i.e. "please fix while in the area") but
MRs that don't change the line will not.

It's not foolproof, but it should be slightly more useful than what we
do right now.