dev.buildfe: Don't fixup JUnit logs
PyTest and Meson both generate JUnit logs, but previously we would tweak
them to represent XFAIL and XPASS as normal FAIL and PASS, respectively.
Unfortunately this also means many MRs have been unnecessarily flagged
with failing tests, due to known XFAILs in the Python unwinding tests.

This is silly, so the tweaks are removed and the JUnit logs generated by
PyTest and Meson are used directly. The normal PASS and FAIL results are
unchanged, but for PyTest:
 - XFAIL is now reported as a SKIPPED test.
 - Non-strict XPASS is still reported as a PASSED test.
 - Strict XPASS is now reported as a FAILED test.
And for Meson:
 - XFAIL and TIMEOUT are now reported as PASSED tests.
 - XPASS is still reported as a PASSED test.

This should result in less annoyance and more green checkmarks.