ci: Redesign handling for images
For a while now, the core idea behind CI pipeline was to (a) get/construct
container image(s) with all the dependencies and then (b) run tests in it.
This became complicated by various attempts to save time by removing parts
of the pipeline (e.g. reuse from the upstream branch instead of constructing
new images), which don't always work *quite* right and can cause surprising
failures.

This commit does away with that approach in favor of a new approach: every
CI pipeline does everything, and caches are used in the jobs to speed up
the jobs individually.

To make this more tractable, the setup sequence is redesigned pretty much
from the ground up. Each container image ("spin") needed is produced by a
heavily customized `spack containerize`. The images are built using buildah,
with access to both the buildah cache and the buildcache. The final jobs
acquire the images *by digest* via dynamic variables.

As a side note, we have 18 spins in the pipeline. That's a lot. We should
triage them a bit in the near future.