ci: Use Spack develop-snapshot images
These snapshots are made weekly and represent a known-working version of Spack.
Using them should reduce the number of random issues we may encounter, especially
if we upstream tests for our use case.

This also increases the usefulness of the Buildah shared layer cache, since layers
can now be (in theory) reused for around a week before invalidation.