ci: Fix `spack install` lint job
The new images have a Spack environment already active, so only one
`hpctoolkit` spec can be added at a time. Adjust to make that work.

Also, the `spack install` job should succeed on the protected branches
since those are in fact published to Spack. Fail the CI pipeline when
the recipe is out-of-date somewhere.