Fx issue 769 (incorrect event name parser)
hpcrun incorrectly parses event name qualifiers if they are prefixed
with numbers

For instance:

hpcrun -e xxx:123b

hpcrun will assume 123b is the period threshold while the event name is
only `xxxx`. This is because we wanted to keep backward compatibility
with a very old hpcrun.

This commit removes backward compatibility.