hpcprof: Use util::File in MetaDB
Unlike the other Sinks, MetaDB wrote the meta.db using the STL std::ofstream
interface instead of the util::File interface. This makes it difficult to
perform wide refactors on the I/O infrastructure, especially for filesystems
that aren't compatible with POSIX.

This commit rewrites MetaDB to use the util::File interface. To make that
easier for this highly structured format, it also adds a new series of
templates designed to provide a RAII/nested-scope style to the writing
process. Overcomplicated but the result is somewhat more straightforward.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/763