dev.buildfe: Also check symbol manifests
hpcrun injects a number of symbol overrides into the application namespace.
We want to reduce our impact on the application, the first step in that
direction is to be aware of what impact we are having now.

This commit extends the `check install` phase (tested in `buildmany`) to
also check the symbol tables of the installed shared libraries. This ensures
our impact is consistent and predictable even as we continue to add
features and perform refactors.