ci: don't use the Spack public buildcache
The Spack public buildcache updates faster than we do, since we only
use the latest weekly `develop` snapshot and not `develop` itself. This
can cause the CI to break when the buildcache format is bumped upstream.

This is a hotfix to unblock CI, but a redesign of our Spack-in-CI setup
is badly needed at this point.