test: enforce presence of test data and fresh targets
The previous version of the tests2/ Meson code intentionally left some
"cracks" in the configuration space, falling into one of these would
permit the configuration to proceed without error but would fail later
in the sequence. A prime example of this is that the fresh-*.tar.xz
targets may not exist if there is some error in the configuration.
This is very confusing and difficult to reason about in CI.

This commit removes this logic, ensuring that a fresh-*.tar.xz target
*always* has the same contents, else the configuration fails. While in
the area, the `fresh-x86-64.tar.xz` tarball has been merged into
`fresh-none.tar.xz` as made possible by recent Dyninst versions.

A similar "crack" allowed the tarballs containing test data to fail to
unpack, resulting in later test failures. This logic has also been
removed, the tarballs unpack or else the configuration fails.

(As a side note, the Meson code is now no longer using disablers. They
are a "poison value" that is difficult to reason about, it's much easier
to understand without using them. You're welcome.)