refactor(dev): Remove "optional" PAPI (--papi=both)
The ./dev script has long supported generating devenvs in a mode which
can support both `--with-papi` and `--without-papi`. This was mostly
used by CI, and was done so that perfmon would not be built separately
from the one bundled with PAPI (i.e. `--without-papi` -> `--with-perfmon`).

This (was always, but now is especially) ludicrous, and complicates
an otherwise simple description of the dependencies. Remove it and just
build a redundant `libpfm4`.

Note that this does not currently affect the build since the Meson code
redundantly prefers PAPI to perfmon. This may change in the future, but
that will be handled in later MRs.