feat(dev): Add ./dev generate-request
This new command compacts (most) possible `./dev generate` calls into a
single command. The result is a machine-readable composition of all the
resulting devenvs in JSON, which can then be used to reconsistute and
install many devenvs based on simple "features" about the base system.

Note that templates can be used with this command, but the output format
only supports a very restricted set of spack.yaml syntax for safety
reasons. Caveat emptor.