ci: Outsource deps images construction
Alright, here's the big upgrade.

The slowest part of the CI pipeline is building the deps images. Making
that faster will complicate this delicate pipeline even further, so
take the "easy" way out: outsource them! Kinda.

The idea here is to separate the container build logic from the rest of
the HPCToolkit CI. Doing this allows the container builds to run on a
separate schedule from the rest of HPCToolkit's CI, removing the
bottleneck in most cases and significantly improving CI performance.

The new images are minimal and don't have any extra packages or
compilers. In the ephemeral job containers we install:
 - The compiler we're compiling with (but no others).
 - file and diffutils. These are implicit dependencies of Autotools, and
   Autotools doesn't handle the error properly when these commands are
   not present(!). Blegh.
 - tar. This is needed for the ext-libs/ copy script.
 - make. This is needed for hpcstruct measurement dir handling.