feat(dev): Add a new ./dev --traceback flag
There are a few cases where an exception is "well-known" and thus
converted into a human-understandable ClickException, which removes the
traceback and other such details. Unfortuately there are occasionally
cases where this pruning removes important information for debugging.

This new flag makes it so the traceback is printed before the exception
is converted, preserving debugging at least to some extent.