build: Remove ext-libs/
`ext-libs/` stores (a number of) copies of our dependencies in our own
install directory. The original intention was to avoid letting the
builder/user experience "dependency hell" when it comes to our library
dependencies. While this works (for now), it is a terrible solution with
many flaws that users are wholly unable to customize.

Spack/Nix/EasyBuild/Homebrew/etc. have "solved" this problem for years
by setting `R(UN)?PATH` in installed binaries, so most users have never
directly experienced this "dependency hell" anyway. The only ones still
experiencing the issues `ext-libs/` ever intended to solve are packagers
(highly experienced in such dependency hellscapes) and developers (us).

This commit removes `ext-libs/` from the install procedure in Autotools.
In practice, this primarily involves deleting less than 50 lines of
Automake and less than 120 lines of dead shell code. The most major
changes are in the `hpcrun` and `hpclink` wrappers, where the paths to
libmonitor.so, libelf.a and libz.a now refer to their install prefixes
instead of `ext-libs/`.

Note that this commit does **NOT** add `R(UN)?PATH`s for all
dependencies. It is expected the builder is aware of this common issue
and will set `LDFLAGS` accordingly to add the paths they see fit.
Spack et al. inject the requisite paths via this and other means,
end-users will not experience any deviation in behavior.

(There are a few dependencies for which `RUNPATH`s are added by the
Autogoo, namely separate CUPTI, IGC and metrics-discovery. These are
retained in this commit, but they should later be reviewed and removed
with heavy testing if not required for correct operation.)

The install manifest checked in CI has been appropriately pruned,
removing over 200 lines of code there.