build: Make final R*PATHs consistent with user flags
The Autogoo for hpcprof and hpcproflm enforce RPATH and add an RPATH for
libstdc++. Both are problematic behaviors, the former because the
builder may exclusively want RUNPATHs, the latter because the C++ may be
libc++ (LLVM) and adding this path will disturb that search order.

This commit removes these behaviors, aligning the installed R*PATHs with
those specified by the builder/user in the LDFLAGS (excepting the
additional paths for the 3 dependencies mentioned in the parent commit).

This commit also removes an empty R*PATH added in hpcrun (the intent was
`$ORIGIN`, but it was not properly quoted so shell expansion made it
empty). With this change, the R*PATHs are byte-for-byte identical across
all installed shared binaries (excepting the deps mentioned previously).

Note that this means builders are expected to add the appropriate
R*PATHs for libstdc++ et al. if needed. In most practical cases, the
builder need not do anything as the packaging of the compiler invisibly
handles this detail (e.g. [Spack-built GCC injects R*PATHs][1]).

[1]: https://github.com/spack/spack/blob/2e9e7ce7c49990cccd8a8d777f0491978d72220d/var/spack/repos/builtin/packages/gcc/package.py#L914-L980