fix(dev): Inject -Wl,-rpath arguments
The builder is now in charge of injecting the appropriate *_link_args to
add R*PATHs to all the (direct) dependencies. This commit extends the
./dev script to generate these paths as part of devenv population.

Unfortunately we can't just query Spack for what RPATHs it would use, so
the logic implemented here is a simplified form that is known to work
if all dependencies are Spack-built. Support for other and more complex
cases will be implemented later as needed.