ci: Don't activate the installed Spack environment
Activating the Spack environment adds the environment's view to the
runtime environment (`PATH` and `LD_LIBRARY_PATH`). This is more
"typical" of a build, but we also want to strongly support "out-of-band"
cases where all the dependencies are built in custom locations... as is
typical in development or cross-arch builds.

This commit deactivates the Spack environment in all CI jobs,
replicating this "out-of-band" case. The only two Spack-built binaries
left on the PATH (`/usr/local/bin`) now are:
 - Python (but not any extensions), needed for the `./dev` script, and
 - Make, needed for `hpcstruct` measurement dir handling.

Note that with this CI does not test the "in-band" case where all
dependencies are in the environment. This case is remarkably similar
to the "out-of-band" case due to how we handle dependencies, so there
isn't much point until https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/709
is implemented.