build: Pass LDFLAGS to the libhpcrun.so link
The final link for libhpcrun.so is done with a manually-crafted command
line, outside of Autotools. This means it does not capture the effects
of LDFLAGS by default. Since we now require the builder to specify
R*PATHs via LDFLAGS, this makes libhpcrun.so unusable in valid cases.

Add LDFLAGS to this command line to ensure R*PATHs appear on
libhpcrun.so in addition to all the other binaries.