fix(hpcrun): Search for ROCProfiler bits at runtime
The released versions of ROCm have thrown us a long list of curveballs.
Most problematic of them come from ROCProfiler which requires that two
environment variables, `ROCP_METRICS` and `HSA_TOOLS_LIB`, are set to
files installed as part of ROCProfiler, `metrics.xml` and
`librocprofiler64.so*` respectively. In this case, our desire for
cross-version compatibility across ROCm versions backfires on us, we
must find these files at runtime instead of at configure-time to
preserve some semblance of consistency.

To make matters worse, these files have changed locations and contents
across ROCm versions and installation modes (unified vs. separate).
There aren't any environment variables consistently set across all the
different ways ROCm can be deployed out in the wild, including
`ROCM_PATH` and `LD_LIBRARY_PATH`.

This commit adds logic to the `hpcrun` wrapper script to search for the
full paths to appropriate files and sets these variables accordingly.
The search starts from the `--rocprofiler-path` passed as an argument,
which defaults to `ROCM_PATH` (fairly common on HPC clusters). This will
not fail if the files cannot be found in reasonable locations, but later
logic inside `libhpcrun.so` will cause the command to crash if the
variables are not set.

Note that this logic currently assumes `librocprofiler64.so.2` is a
can be used in place of `librocprofiler64.so.1` if the latter is not
present. This has API/ABI consequences and will only work if later ROCm
versions place backwards-incompatible API/ABI changes in a file named
something else, like `librocprofiler64.so.3`.
Hopefully they don't. Something to fix another day.