refactor(dev): Remove AutogenEnv in favor of meson setup
The "autogen" environment was original written to add a way to
regenerate the Autogoo with exactly the right versions of the various
Autotools. But this effort is now duplicated during a Meson-ic build,
Meson will run `./autogen` with the appropriate tool versions as part
of the initial `meson setup`.

This commit removes AutogenEnv, leaving behind DevEnv as the only class
of SpackEnv. All commands and options that used AutogenEnv have been
removed, and the CI now performs its `autoreconf` check by running
`./dev populate` in place of `./dev autogen`.

BREAKING CHANGE: `./dev autogen` and is no longer a subcommand and
`./dev ... --autogen` flags have been removed.