build(dev): Add dependency on Spiqa
[Spiqa](https://gitlab.com/hpctoolkit/spiqa) is the new Python library
to interface with (parts of) Spack. It contains interfaces that need to
be kept consistent across the various HPCToolkit repositories. We should
use in the `./dev` script in place of various Spack logics.

This commit just adds Spiqa as a dependency in the pyproject.toml.
Later commits will do the actual refactors.