ci: Extend the ROCm testing matrix
Add new CI jobs to test the full breadth of ROCm versions. Now all ROCm
versions 5.1-5.6 are tested both in `buildmany` and `check` capacities.

The current behavior of `hpcrun` requires that [the ROCm libraries are
on the library search path][1]. This is not true in the ROCm 5.1
container, and only is true in later versions due to a poor packaging
decision on the AMD side. Thus, the `check` jobs for ROCm extend the
`LD_LIBRARY_PATH` with the ROCm being tested.

[1]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/784