build(meson): Add missing inputs to ./configure
There are a number of human-written files that `./configure` reads that
`./autogen` does not. Since these are outside of Meson's tracking,
altering these files would not trigger a Meson reconfiguration, which
causes confusion and other consternations.

Add these files to the command line for `./autogen` to include them in
Meson's tracking and remove this issue.