fix(hpcrun): Gracefully handle no ROCProfiler for -L
If `ROCP_METRICS` and `HSA_TOOLS_LIB` are unset or empty, we know that
the wrapper script was unable to find the bits for ROCProfiler. In this
case `hpcrun -L` crashes attempting to bind and load ROCProfiler.

Avoid the crash by checking that these two variables are set and
non-empty before attempting to bind ROCProfiler. Instead, print a
message noting that ROCProfiler is not available and a terse instruction
on where to look to fix the problem.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/784