refactor: Remove doc/www
This directory contained the code for the http://hpctoolkit.org website.
None of this website actually documents HPCToolkit, instead deferring to
the manual (`doc/HPCtoolkit-users-manual.pdf`) or to the man pages
(`doc/man`). It also contains an annoyingly large amount of committed
generated code, making it a tangled mess to maintain.

The necessary code has been extracted to a dedicated repository
(https://gitlab.com/hpctoolkit/hpctoolkit.org), so the copy in this repo
can be safely deleted. As such the website is no longer installed to
`<prefix>/share/doc/hpctoolkit` as part of `make install`.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/715