ci: Prune the CI test matrix
We test a lot of things in CI, but many of the jobs are there to fill
out the matrix of possibilities and do not provide much benefit to run
every. single. time. So we purge the ones that are unlikely to provide
us with much effective gain in the long run, especially given their
overall time cost.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/766
Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/786