fix(hpcprof): Correct inconsistent logic for empty FlowGraphs
A FlowGraph may have no Templates, but if it has a handler `hpcprof` crashes with an assertion
failure. This is unnecessary, and causes crashes in practice when StructFile is unsuccessful at
finding a valid path to the kernel entry point (i.e. incomplete call graph from `hpcstruct`).

Adjust the assertion to allow this case, and adjust any code that assumed `!m_handler` implied
`m_templates.empty()`.