ci: Remove usage of ./dev buildfe
Rewrite the CI job scripts to use shell scripting instead of the now
obsolete `./dev buildfe` command. The log output now makes it much
clearer exactly what is being done when it is being done, at the cost
of a slightly less "pretty" output. Oh well.

The reproduction commands are now part of the jobs' `after_script:`
commands. This is much cleaner as the CI configuration is what
determines the environment for a job, not the `./dev` script.

The buildmany matrix is much more verbose now to keep the advanced shell
scripting to a minimum.