fix(dev): Correct the install manifest
The install manifest was reporting a number of files as "unexpected"
when they really weren't, namely a number of `ext-libs/` and a few new
files written since the install manifest was first written.

I don't know why these didn't cause warnings back with `builefe`...
maybe there was a flaw in the reporting system?