fix(tests): Skip tests that require submodules if not cloned
Parts of the test suite use data that is not included in this repository,
instead using a separable submodule. This data is large and not
necessary unless you plan to run tests, but Meson configuration fails if
it is not present.

This commit tweaks the logic so that if these files seem to be missing,
it simply skips the corresponding tests.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/803