fix(hpcprof): Elide mutual recursion from GPU Context reconstruction
Some GPU binaries contain complex recursion between functions. Prior to
this commit, this would end up deadlocking `hpcprof` because it
recursively resolves the Templates for every FlowGraph mapping to a
particular address.

This commit adds some additional postprocessing on the call graph
reported by `hpcstruct` to "fudge" the call graph into a DAG by
"summarizing away" possible recursion. In short, if `A -> B <-> C -> D`,
we summarize that as just `A -> C -> D` and `A -> B`.

Note that this "fudge" is a simplistic solution, although the graph is
semi-reasonable the weights derived from it may not be in particularly
complex cases. Since these weights are a whole pile of fudges anyway,
we don't emit any special warning for this particular fudge.
Caveat emptor.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/800