docs: Update README.md with Meson build instructions
The preferred method moving forward for developer builds should be
Meson-based. Now that the `./dev` script is no longer a necessity to
achieve this, we can simplify the build instructions to the traditional
"`meson setup/compile/install`," modulo Spack and the weird way of
specifying dependencies (prefix instead of pkg-config/cmake/etc. flags).

Like any good "getting started" documentation, put in front and center
in the README.md instead of linking to the web site.

This new documentation expects the reader to be familiar with Spack and
Meson (or at least `configure && make && make install`-type builds).
For new readers, much of the text links out to the relevant parts of the
official documentation for those tools to ease navigation.

For completeness, the most important fragment of the Autotools-based
build instructions is included in a separate section of the README.md.