tests: Refactor Python out into "hpctesttool"
Previously, all test scripts are written in Python and use a series of
Python utilities in a side directory. This side directory depends on
packages from the PyPI and uses Python packaging tools (Poetry). The
"proper" way to check for dependencies and such is to install it in
a Python virtual environment, but managing that gets overly complicated.

Most tests boil down to "run `hpc* ...` and then check the output," so
Python is only really needed for the second half of that statement.
This commit makes the tests clearer and simpler by rewriting them into
(short!) shell scripts that call `hpctesttool`, a new (uninstalled)
utility that is effectively the Python halves of the tests.

Tearing this apart significantly simplifies the process of managing and
running tests. The Python scripts do not need to be pre-processed to
rewrite the shebangs, `PYTHONPATH` does not need to be fiddled with,
and we don't have nonstandard `HPCTOOLKIT_APP_*` variables running
around. This also means Meson does not need to reconfigure every time
a test script is edited, which is a significant DevX boon.

I'm not a fan of using shell scripts here, but I have gone this route
because the better alternative (Python's `subprocess`) requires more
lines of code and is overall less familiar. The shell scripts are short
and simple in an effort to avoid the usual issues with using shell.