build(hpctesttool): Vendor all dependencies
The `hpctesttool` has a number of dependencies on PyPI packages. At the
moment these packages must be installed in the `python3` passed to Meson,
but since `hpctesttool` is a Python package itself it must be installed
into a virtual environment (venv) before it can be used. The end result
is that users must install suitable versions of `click`, `pyparsing`,
`pytest` and `ruamel.yaml` into their system or user installations.
While users can do this, these installations are not isolated which
risks a kind of dependency hell, one venvs were meant to solve.

In the Python packaging world the typical solution is to just let `pip`
install the dependencies it needs. Unfortunately this is not suitable
for our case, since we want the `meson setup/compile/install` process to
avoid accessing the internet.

There are two solutions to this problem. One pushes the responsibility
onto the user to install these packages, in a venv if necessary. This
commit implements the other solution: vendor all the dependencies.
The user need only provide a Python 3.10+ and we handle the rest.

This commit adds 1.2MB of wheels in `lib/python/wheels`. At build time,
these files are installed to a venv in the build directory, which then
provides the `hpctesttool` for the test suite. Keeping things consistent,
the exact versions vendored are generated from the `poetry.lock` file.

As an exception, `ruamel.yaml.clib` is swapped out for an empty "shim"
package. This is an optional dependency for `ruamel.yaml`, but without
this shim Pip will refuse to install or worse, attempt to build the C
extension from source itself.