fix(hpcstruct): Burn in path to nvdisasm
Currently `hpcstruct --gpucfg=yes` will always use the first `nvdisasm`
found on the PATH, falling back to `--gpucfg=no` behavior if it cannot
be found. This makes `hpcstruct` behave erratically depending on the
current PATH and/or loaded modules ([#799]), causing issues for both
users and the test suite ([#798]).

This commit alters `hpcstruct` to *always* use the `nvdisasm` that it
was compiled with. If `hpcstruct` was not compiled with CUDA support,
it will fall back to `--gpucfg=no` behavior. In effect, this change
makes `nvdisasm` a library-style dependency of `hpcstruct`, thus
handily removing it from the user's consciousness and improving the UX.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/801
Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/798

[#799]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/799
[#798]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/798