fix(hpcstruct): Make --gpucfg=yes fatal if GPU support not compiled in
If you request a GPU CFG (`hpcstruct --gpucfg=yes`) for a file that
`hpcstruct` is not prepared to generate CFGs for (e.g. `~cuda` build),
then `hpcstruct` will happily and quietly produce results as if you had
instead requested `--gpucfg=no`. This is surprising and confusing from
a user's perspective, and difficult to identify when it does occur.

This commit alters the semantics of `--gpucfg=yes`, now if `hpcstruct`
can't provide the CFGs it will error out instead of falling back to the
behavior of `--gpucfg=no`. Now that `nvdisam` is treated as a library
dependency, this behavior is easy to predict for users.
Need GPU CFGs? Compile `+cuda`. Simple and concise.

To retain this behavior for measurement directories, the Makefile is
adjusted to expose a non-zero exit code if any of the inner `hpcstruct`
calls fails (previously any errors were aborbed). The overall process
will still continue even if a binary fails, this only alters the exit
code after all binaries have (attempted) processing.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/799